First commit for libxml2_7

+++ testlimits.c
@@ -1284,14 +1284,13 @@ saxTest(const char *filename, size_t limit, int options, int fail) {
         if (fail)
             res = 0;
         else {
-            fprintf(stderr, "Failed to parse '%s' %lu\n", filename,
-                    (unsigned long) limit);
+            fprintf(stderr, "Failed to parse '%s' %lu\n", filename, limit);
             res = 1;
         }
     } else {
         if (fail) {
             fprintf(stderr, "Failed to get failure for '%s' %lu\n",
-                    filename, (unsigned long) limit);
+                    filename, limit);
             res = 1;
         } else
             res = 0;
@@ -1340,7 +1339,7 @@ readerTest(const char *filename, size_t limit, int options, int fail) {
                         filename, crazy_indx);
             else
                 fprintf(stderr, "Failed to parse '%s' %lu\n",
-                        filename, (unsigned long) limit);
+                        filename, limit);
             res = 1;
         }
     } else {
@@ -1350,7 +1349,7 @@ readerTest(const char *filename, size_t limit, int options, int fail) {
                         filename, crazy_indx);
             else
                 fprintf(stderr, "Failed to get failure for '%s' %lu\n",
-                        filename, (unsigned long) limit);
+                        filename, limit);
             res = 1;
         } else
             res = 0;
